requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pyaes==1.6.1
pyasn1==0.5.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
Telethon==1.33.1
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
